Additional code refactor related to #7124 to avoid code duplication
