+++ core/src/main/java/org/fao/geonet/kernel/setting/SettingManager.java
@@ -1,5 +1,5 @@
 //=============================================================================
-//===	Copyright (C) 2001-2021 Food and Agriculture Organization of the
+//===	Copyright (C) 2001-2023 Food and Agriculture Organization of the
 //===	United Nations (FAO-UN), United Nations World Food Programme (WFP)
 //===	and United Nations Environment Programme (UNEP)
 //===
@@ -457,9 +457,36 @@ public class SettingManager {
     String getServerURL() {
         String protocol = getValue(Settings.SYSTEM_SERVER_PROTOCOL);
         String host = getValue(Settings.SYSTEM_SERVER_HOST);
-        String port = getValue(Settings.SYSTEM_SERVER_PORT);
+        Integer port = getServerPort();
 
-        return protocol + "://" + host + (isPortRequired(protocol, port) ? ":" + port : "");
+        StringBuffer sb = new StringBuffer(protocol + "://");
+
+        sb.append(host);
+
+        if (isPortRequired(protocol, port + "")) {
+            sb.append(":");
+            sb.append(port);
+        }
+
+        return sb.toString();
+    }
+
+    public Integer getServerPort() {
+        String protocol = getValue(Settings.SYSTEM_SERVER_PROTOCOL);
+
+        // some conditional logic to handle the case where there's no port in the settings
+        Integer sitePort;
+
+        Integer configuredPort = toIntOrNull(Settings.SYSTEM_SERVER_PORT);
+        if (configuredPort != null) {
+            sitePort = configuredPort;
+        } else if (protocol.equalsIgnoreCase(Geonet.HttpProtocol.HTTPS)) {
+            sitePort = Geonet.DefaultHttpPort.HTTPS;
+        } else {
+            sitePort = Geonet.DefaultHttpPort.HTTP;
+        }
+
+        return sitePort;
     }
 
     public static boolean isPortRequired(String protocol, String port) {
@@ -469,6 +496,14 @@ public class SettingManager {
             return false;
         } else {
             return true;
+        }
+    }
+
+    private Integer toIntOrNull(String key) {
+        try {
+            return Integer.parseInt(getValue(key));
+        } catch (NumberFormatException e) {
+            return null;
         }
     }
 }
